Dissolve cpgqueryingtests.
