--- cpgqueryingtests/build.sbt
@@ -1,13 +1,0 @@
-name := "cpgqueryingtests"
-
-dependsOn(
-  Projects.semanticcpg,
-  Projects.dataflowengine % Test
-)
-
-libraryDependencies ++= Seq(
-  "org.scalatest" %% "scalatest" % "3.0.3" % Test,
-  "io.shiftleft" %% "fuzzyc2cpg" % Versions.fuzzyc2cpg
-)
-
-scalacOptions ++= Seq("-deprecation", "-feature")
