+++ semanticcpg/src/test/scala/io/shiftleft/queryprimitives/steps/LanguageFrontend.scala
@@ -1,6 +1,7 @@
-package io.shiftleft.cpgqueryingtests.codepropertygraph
+package io.shiftleft.queryprimitives.steps
 
 import java.io.File
+
 import io.shiftleft.fuzzyc2cpg.FuzzyC2Cpg
 import io.shiftleft.fuzzyc2cpg.output.protobuf.OutputModuleFactory
