+++ semanticcpg/src/test/scala/io/shiftleft/queryprimitives/steps/MethodTests.scala
@@ -1,9 +1,7 @@
-package io.shiftleft.cpgqueryingtests.steps
+package io.shiftleft.queryprimitives.steps
 
 import io.shiftleft.codepropertygraph.generated.nodes
 import org.scalatest.{Matchers, WordSpec}
-
-import io.shiftleft.cpgqueryingtests.codepropertygraph.CpgTestFixture
 
 class MethodTests extends WordSpec with Matchers {
